Update README.md
This project is implemented in Python. 
Used A* algorithm to find the shortest path between two points. 
Used python libraries like pygame, pyglet for visualizing the path.

What is A* Search Algorithm?
 A* Search algorithm is one of the best and popular techniques used in path-finding and graph traversals.

Why A* Search Algorithm? 
Informally speaking, A* Search algorithms, unlike other traversal techniques, it has “brains”. What it means is that it is really a smart algorithm that separates it from the other conventional algorithms. This fact is cleared in detail in the below sections. And it is also worth mentioning that many games and web-based maps use this algorithm to find the shortest path very efficiently (approximation).